string-1 assignment commit

+++ Week6_Array2D/Assignment-1_2D/fivthAns.cpp
@@ -1,11 +1,11 @@
 // Write a program to print the row number having the maximum sum in a given matrix.
 #include<iostream>
+#include<climits>
 using namespace std;
 int main (){
 int arr[3][4] ={1, 3, 5, 7, 3, 4, 7, 8,1, 4, 12, 3};
-int Maxrowsum =INT8_MIN;
+int Maxrowsum =INT_MIN;
 int maxRow = -1;
-
 for(int i=0; i<3; i++){
     int rowsum =0;
     for(int j=0; j<4; j++){
@@ -20,5 +20,6 @@ for(int i=0; i<3; i++){
 }
 
 cout<<maxRow+1;
+
 
 }
